adds edit and create

+++ routes/projects.js
@@ -1,5 +1,6 @@
 const router = require('express').Router();
 const Project = require('../models/Project');
+const ErrorResponse = require('../utils/error');
 
 // @desc    Get all projects
 // @route   GET /
@@ -7,11 +8,10 @@ const Project = require('../models/Project');
 router.get('/', async (req, res, next) => {
   try {
     const projects = await Project.find({});
-    if (projects.length === 0) {
-      res.status(200).json({ response: 'No projects found in the database ' });
-    } else {
-      res.status(200).json({ data: projects })
+    if (!projects) {
+      next(new ErrorResponse('No projects found', 404));
     }
+    res.status(200).json({ data: projects })
   } catch (error) {
     next(error);
   }
@@ -25,10 +25,9 @@ router.get('/:id', async (req, res, next) => {
   try {
     const project = await Project.findById(id);
     if (!project) {
-      res.status(404).json({ response: 'Project not found' });
-    } else {
-      res.status(200).json({ data: project })
+      next(new ErrorResponse(`Project not found by id: ${id}`, 404));
     }
+    res.status(200).json({ data: project })
   } catch (error) {
     next(error);
   }
@@ -41,6 +40,9 @@ router.post('/', async (req, res, next) => {
   const { title, description } = req.body;
   try {
     const project = await Project.create({ title, description });
+    if (!project) {
+      next(new ErrorResponse('An error ocurred while creating the project', 500));
+    }
     res.status(201).json({ data: project })
   } catch (error) {
     next(error);
@@ -54,8 +56,13 @@ router.put('/:id', async (req, res, next) => {
   const { id } = req.params;
   const { title, description } = req.body;
   try {
-    const updatedProject = await Project.findByIdAndUpdate(id, { title, description }, { new: true });
-    res.status(202).json({ data: updatedProject })
+    const project = await Project.findById(id);
+    if (!project) {
+      next(new ErrorResponse(`Project not found by id: ${id}`, 404));
+    } else {
+      const updatedProject = await Project.findByIdAndUpdate(id, { title, description }, { new: true });
+      res.status(202).json({ data: updatedProject })
+    }
   } catch (error) {
     next(error);
   }
@@ -67,8 +74,13 @@ router.put('/:id', async (req, res, next) => {
 router.delete('/:id', async (req, res, next) => {
   const { id } = req.params;
   try {
-    const deleted = await Project.findByIdAndDelete(id);
-    res.status(202).json({ data: deleted });
+    const project = await Project.findById(id);
+    if (!project) {
+      next(new ErrorResponse(`Project not found by id: ${id}`, 404));
+    } else {
+      const deleted = await Project.findByIdAndDelete(id);
+      res.status(202).json({ data: deleted });
+    }
   } catch (error) {
     next(error);
   }
